점수 소수 부분에 오차가 발생하는 문제 수정, 인트로 모드 추가

+++ GL2D/ImageUtil.h
@@ -13,7 +13,7 @@ public:
 	void Init();
 	void LoadImageFromList();
 	unsigned int SetImage(std::string ImageName);
-	GLfloat Aspect(int Width, int Height);
+	GLfloat Aspect(float Width, float Height);
 	void Draw(unsigned int ImageVar);
 };
 extern ImageUtil imageUtil;
